data-ingestion-service/src/main/java/gov/cdc/dataingestion/share/ControllerExceptionHandler.java
package gov.cdc.dataingestion.share;

import gov.cdc.dataingestion.share.model.ErrorResponse;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.server.ResponseStatusException;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

@ControllerAdvice
/**
 1118 - require constructor complaint
 125 - comment complaint
 6126 - String block complaint
 1135 - todos complaint
 * */
@SuppressWarnings({"java:S1118","java:S125", "java:S6126", "java:S1135"})
public class ControllerExceptionHandler extends ResponseEntityExceptionHandler {
    @ExceptionHandler({Exception.class, ResponseStatusException.class})
    public ResponseEntity<ErrorResponse> handleException(Exception ex) {
        ErrorResponse errorResponse = new ErrorResponse();
        errorResponse.setStatusCode(HttpStatus.INTERNAL_SERVER_ERROR.value());
        errorResponse.setMessage("An internal server error occurred.");

        String errMessage;
        if (ex instanceof ResponseStatusException responseStatusException) {
            errMessage = responseStatusException.getReason();
        } else {
            errMessage = ex.getMessage();
        }
        errorResponse.setDetails(errMessage);

        return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
    }


}
